Update map structure to function better in pure react (#187)
