+++ ui/src/components/Map/index.tsx
@@ -3,16 +3,35 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { lazy, Suspense } from 'react';
-import { MapComponentProps } from '@/components/Map/types';
+import React, { useEffect, useRef } from 'react';
+import MapboxDraw from '@mapbox/mapbox-gl-draw';
+import MapboxGeocoder from '@mapbox/mapbox-gl-geocoder';
+import mapboxgl, { Map as _Map, IControl } from 'mapbox-gl';
+import { MapComponentProps, MapEventKey } from '@/components/Map/types';
+import {
+  addClickFunctions,
+  addControls,
+  addCustomControls,
+  addHoverFunctions,
+  addLayers,
+  addMouseMoveFunctions,
+  addSources,
+} from '@/components/Map/utils';
+import { useMap } from '@/contexts/MapContexts';
 
-// Dynamic load causes a double render even without strict mode
-const ClientSideMap = lazy(() => import('./ClientSide'));
+import 'mapbox-gl/dist/mapbox-gl.css';
+import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
+import '@mapbox/mapbox-gl-draw/dist/mapbox-gl-draw.css';
+
+import FeatureService, { FeatureServiceOptions } from '@hansdo/mapbox-gl-arcgis-featureserver';
+import { createRoot } from 'react-dom/client';
+
+FeatureService.prototype._setAttribution = function () {
+  // Stub to prevent attribution bug
+};
 
 /**
- * This component renders the map component using a lazy load.
- *
- * The map component initializes and renders a Mapbox GL map with specified sources, layers, and controls.
+ * This component initializes and renders a Mapbox GL map with specified sources, layers, and controls.
  * It handles map loading, style changes, and cleanup on component unmount. Once the map is initialized
  * it the map object is stored into the map context provider to allow referencing across the application.
  *
@@ -33,14 +52,163 @@ const ClientSideMap = lazy(() => import('./ClientSide'));
  *
  * @component
  */
-const Map: React.FC<MapComponentProps> = (props) => {
+const MapComponent: React.FC<MapComponentProps> = (props) => {
+  const {
+    id,
+    sources,
+    layers,
+    options,
+    controls,
+    customControls,
+    accessToken,
+    persist = false,
+    geocoder,
+    draw,
+    eventHandlers = {},
+  } = props;
+
+  const mapContainerRef = useRef<HTMLDivElement | null>(null);
+
+  // Prevent duplicate map instantiations in dev mode
+  const isFirstLoad = useRef(true);
+  const {
+    map,
+    hoverPopup,
+    persistentPopup,
+    draw: drawInstance,
+    root,
+    container,
+    setMap,
+  } = useMap(id);
+
+  useEffect(() => {
+    if (!isFirstLoad.current) {
+      return;
+    }
+    isFirstLoad.current = false;
+
+    if (!map && mapContainerRef.current) {
+      mapboxgl.accessToken = accessToken;
+      const newMap = new mapboxgl.Map({
+        ...options,
+        container: mapContainerRef.current,
+      });
+      const hoverPopup = new mapboxgl.Popup({
+        closeButton: false,
+        closeOnClick: false,
+      });
+
+      const persistentPopup = new mapboxgl.Popup();
+
+      let _geocoder: MapboxGeocoder | null = null;
+      if (geocoder) {
+        const { position, ...geocoderWithoutPosition } = geocoder;
+        _geocoder = new MapboxGeocoder({
+          accessToken: mapboxgl.accessToken,
+          ...geocoderWithoutPosition,
+        });
+        if (position) {
+          newMap.addControl(_geocoder, position);
+        }
+      }
+      let _draw: MapboxDraw | null = null;
+      if (draw) {
+        const { position, ...drawWithoutPosition } = draw;
+        _draw = new MapboxDraw(drawWithoutPosition);
+        if (position) {
+          newMap.addControl(_draw as unknown as IControl, position);
+        } else {
+          newMap.addControl(_draw as unknown as IControl);
+        }
+      }
+
+      const container = document.createElement('div');
+      container.setAttribute('id', 'customPopupContent');
+      const root = createRoot(container);
+
+      Object.entries(eventHandlers).forEach(([event, value]) => {
+        const key = event as MapEventKey;
+        if (event in newMap && typeof newMap[key]?.enable === 'function') {
+          value ? newMap[key].enable() : newMap[key].disable();
+        } else {
+          console.error('Invalid event key used: ', key);
+        }
+      });
+
+      newMap.once('load', () => {
+        const createFeatureService = (
+          sourceId: string,
+          map: _Map,
+          options: FeatureServiceOptions
+        ) => new FeatureService(sourceId, map, options);
+
+        setMap(newMap, hoverPopup, persistentPopup, _geocoder, _draw, root, container);
+        addSources(newMap, sources, createFeatureService);
+        addLayers(newMap, layers);
+        addHoverFunctions(newMap, layers, hoverPopup, persistentPopup, _draw, root, container);
+        addClickFunctions(newMap, layers, hoverPopup, persistentPopup, _draw, root, container);
+        addMouseMoveFunctions(newMap, layers, hoverPopup, persistentPopup, _draw, root, container);
+        addControls(newMap, controls);
+        addCustomControls(newMap, customControls);
+      });
+    } else if (
+      persist &&
+      map &&
+      mapContainerRef.current &&
+      mapContainerRef.current.innerHTML.length === 0
+    ) {
+      const container = map.getContainer();
+      const parent = container.parentNode;
+
+      if (mapContainerRef.current && container !== mapContainerRef.current) {
+        if (parent) {
+          parent.removeChild(container);
+        }
+        mapContainerRef.current.appendChild(container);
+        map.resize();
+      }
+    }
+
+    return () => {
+      if (!persist && map) {
+        map.remove();
+      }
+      if (!persist && root) {
+        root.unmount();
+      }
+    };
+  }, []);
+
+  useEffect(() => {
+    if (!map || !hoverPopup || !persistentPopup || !root || !container) {
+      return;
+    }
+
+    map.on('style.load', () => {
+      const createFeatureService = (sourceId: string, map: _Map, options: FeatureServiceOptions) =>
+        new FeatureService(sourceId, map, options);
+
+      // Layers reset on style changes
+      addSources(map, sources, createFeatureService);
+      addLayers(map, layers);
+      addHoverFunctions(map, layers, hoverPopup, persistentPopup, drawInstance, root, container);
+      addClickFunctions(map, layers, hoverPopup, persistentPopup, drawInstance, root, container);
+      addMouseMoveFunctions(
+        map,
+        layers,
+        hoverPopup,
+        persistentPopup,
+        drawInstance,
+        root,
+        container
+      );
+    });
+  }, [map]);
+
+  // Style the container using #map-container-${id} in a global css file
   return (
-    <>
-      <Suspense>
-        <ClientSideMap {...props} />
-      </Suspense>
-    </>
+    <div data-testid={`map-container-${id}`} id={`map-container-${id}`} ref={mapContainerRef} />
   );
 };
 
-export default Map;
+export default MapComponent;
